modify cleaning processing of text

+++ preprocessing_data/create_index_laberta.py
@@ -25,34 +25,39 @@ folder_path = '/work/pnrr_itserr/WP4-embeddings/latin_data/db_data'
 output_path = '/work/pnrr_itserr/WP4-embeddings/index_path'
 # output_path = '/work/pnrr_itserr/WP4-embeddings/index_path/debug'
 
-debug_for = os.listdir(folder_path) #[:10]
-for el_auth in tqdm(debug_for, mininterval=1, maxinterval=len(os.listdir(folder_path))):
-    el_auth_path = os.path.join(folder_path, el_auth)
-    for el_sample in os.listdir(el_auth_path):
-        el_sample_path = os.path.join(el_auth_path, el_sample)
-        with open(el_sample_path, 'r') as f:
-            data = json.load(f)
-        for i in tqdm(range(0, len(data['content']), batch_size)):
-                        
-            with torch.no_grad():
-                val = data['content'][i:i+batch_size]
-                tokenized = tokenizer(val, padding=True, truncation=True, return_tensors='pt').to(model.device)
-                outputs = model(**tokenized)
-                model_cls = outputs.last_hidden_state[:, 0]
-                model_cls = model_cls.cpu().detach().numpy()
-                faiss.normalize_L2(model_cls)
-                faiss_list_cls.append(model_cls)
-                
-            for j in range(0, len(val), 1):
-                faiss_list_id.append(data['id'] + '_' + str(j))
-
-faiss_list_cls = np.concatenate(faiss_list_cls, axis=0)
-
 tmp_save_directory = os.path.join(output_path, 'tmp_embed_cls.npy')
-np.save(tmp_save_directory, faiss_list_cls)
-autofaiss.build_index(embeddings=output_path, index_path=output_path+"/knn.index", index_infos_path=output_path+"/index_infos.json")
+if not os.path.exists(tmp_save_directory):
+    print("Creating index")
+    debug_for = os.listdir(folder_path) #[:10]
+    for el_auth in tqdm(debug_for, mininterval=1, maxinterval=len(os.listdir(folder_path))):
+        el_auth_path = os.path.join(folder_path, el_auth)
+        for el_sample in os.listdir(el_auth_path):
+            el_sample_path = os.path.join(el_auth_path, el_sample)
+            with open(el_sample_path, 'r') as f:
+                data = json.load(f)
+            for i in range(0, len(data['content']), batch_size):
+                            
+                with torch.no_grad():
+                    val = data['content'][i:i+batch_size]
+                    tokenized = tokenizer(val, padding=True, truncation=True, return_tensors='pt').to(model.device)
+                    outputs = model(**tokenized)
+                    model_cls = outputs.last_hidden_state[:, 0]
+                    model_cls = model_cls.cpu().detach().numpy()
+                    faiss.normalize_L2(model_cls)
+                    faiss_list_cls.append(model_cls)
+                    
+                for j in range(0, len(val), 1):
+                    faiss_list_id.append(data['id'] + '_' + str(j))
 
-with open(os.path.join(output_path, 'knn.json'), 'w') as f:
-    json.dump(faiss_list_id, f)
+    faiss_list_cls = np.concatenate(faiss_list_cls, axis=0)
+    np.save(tmp_save_directory, faiss_list_cls)
+    
+    with open(os.path.join(output_path, 'knn.json'), 'w') as f:
+        json.dump(faiss_list_id, f)
+
+else:
+    print("Loading index")
+    faiss_list_cls = np.load(tmp_save_directory)
+autofaiss.build_index(embeddings=output_path, index_path=output_path+"/knn.index", index_infos_path=output_path+"/index_infos.json")
     
 print('Done')
